[wasm][debugger] Truncate protocol log messages to 64k in length
